parserService/requirements.txt
# ORM and database
SQLAlchemy>=2.0,<2.1
asyncpg>=0.29
alembic>=1.13

# Messaging
aio-pika>=9.4.0

# Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv==1.0.1
setuptools
